bilibili-live-monitor-django/requirements.txt
Django>=4.2.0
redis>=4.5.0
bilibili-api-python>=16.0.0
requests>=2.28.0
aiohttp>=3.8.0
websockets>=11.0.0